# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  # Do not change the title, if the title changes, the import paths will be broken
  version: 0.2.0
  description: AdsGram Telegram Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register or login a Telegram user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "200":
          description: User registered or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{telegramId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile and balance
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tasks/{telegramId}:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all tasks with completion status for a user
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksResponse"
...
  /mini/tasks/{telegramId}:
    get:
      operationId: getMiniTasks
      tags: [mini-tasks]
      summary: List all tasks for the mini-app
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiniTaskList"
...
  /mini/tasks/{id}/complete:
    post:
      operationId: completeMiniTask
      tags: [mini-tasks]
      summary: Complete a task and claim reward
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MiniTaskCompleteInput"
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiniTaskCompleteResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        cooldownSeconds:
          type: integer
      required: [error]
    RegisterUserRequest:
      type: object
      properties:
        telegramId:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        photoUrl:
          type: string
        referredBy:
          type: string
        deviceId:
          type: string
      required: [telegramId]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: string
        firstName:
          type: string
        photoUrl:
          type: string
        coins:
          type: integer
        ton:
          type: number
        tonyxCoins:
          type: integer
        boostRate:
          type: number
        totalAdsWatched:
          type: integer
        totalReferrals:
          type: integer
        isBlocked:
          type: boolean
        isAdmin:
          type: boolean
        depositCode:
          type: string
        warningCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, telegramId, coins, ton, tonyxCoins, boostRate, totalAdsWatched, totalReferrals, isBlocked, isAdmin, createdAt]
...
    TaskItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        reward:
          type: integer
        type:
          type: string
          enum: [subscribe, visit, watch_ads, invite_friends]
        requirement:
          type: integer
        link:
          type: string
        completed:
          type: boolean
        completedAt:
          type: string
          format: date-time
      required: [id, title, description, reward, type, requirement, completed]
    TasksResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskItem"
      required: [tasks]
...
    MiniTask:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
        link:
          type: ["string", "null"]
        reward:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, title, type, reward, completed, createdAt]
    MiniTaskList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/MiniTask"
      required: [tasks]
...
    MiniTaskCompleteInput:
      type: object
      properties:
        telegramId:
          type: string
      required: [telegramId]
    MiniTaskCompleteResult:
      type: object
      properties:
        tonyxEarned:
          type: integer
        tonEarned:
          type: number
        newBalance:
          type: integer
        newTon:
          type: number
        taskTitle:
          type: string
      required: [tonyxEarned, newBalance, taskTitle]
...