# lib/api-zod/src/generated/types/miniTaskCompleteResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AdsGram Telegram Bot API
 * OpenAPI spec version: 0.2.0
 */

export interface MiniTaskCompleteResult {
  tonyxEarned: number;
  tonEarned?: number;
  newBalance: number;
  newTon?: number;
  taskTitle: string;
}